requirements.txt
appnope==0.1.3
APScheduler==3.10.1
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.1
black==23.7.0
bs4==0.0.1
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
coverage==7.2.3
decorator==5.1.1
distlib==0.3.7
Django==4.1
django-apscheduler==0.6.2
django-extensions==3.2.3
exceptiongroup==1.1.1
executing==1.2.0
factory-boy==3.3.0
Faker==19.3.0
feedparser==6.0.8
filelock==3.12.2
flake8==6.1.0
identify==2.5.26
idna==3.4
iniconfig==2.0.0
install==1.3.5
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
pluggy==1.0.0
pre-commit==3.3.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.14.0
pytest==7.3.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.31.0
sentry-sdk==1.29.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.4
virtualenv==20.24.2
wcwidth==0.2.6
